Getting raw photo from server. Allows to zoom it and double tap to return to default zoom.

+++ UnsplashPhotoEditor/Models/Photo.swift
@@ -1,12 +1,12 @@
-import Foundation
+import UIKit
 
 class Photo: Codable {
     let id: String
     let urls: PhotoUrls
-    let width: Int?
-    let height: Int?
+    let width: CGFloat?
+    let height: CGFloat?
     
-    init?(_ id: String?, urls: PhotoUrls?, width: Int?, height: Int?) {
+    init?(_ id: String?, urls: PhotoUrls?, width: CGFloat?, height: CGFloat?) {
         guard let id = id, let photoUrls = urls else { return nil }
         self.id = id
         self.urls = photoUrls
